Adding more index file entries and cleaning up world.ts imports
