+++ src/game-engine/net/inbound-packets/npc-interaction-packet.ts
@@ -1,5 +1,5 @@
 import { world } from '@engine/game-server';
-import { World } from '@engine/world';
+import { World } from '@engine/world/world';
 import { logger } from '@runejs/core';
 import { Player } from '@engine/world/actor/player/player';
 import { PacketData } from '@engine/net/inbound-packet-handler';
