+++ src/game-engine/net/inbound-packets/player-interaction-packet.js
@@ -1,6 +1,6 @@
 import { playerOptions } from '../../world/actor/player/player';
 import { world } from '../../game-server';
-import { World } from '../../world';
+import { World } from '../../world/world';
 import { logger } from '@runejs/core';
 
 const playerInteractionPacket = (player, packet) => {
