+++ src/game-engine/net/outbound-packet-handler.ts
@@ -7,7 +7,7 @@ import { Position } from '@engine/world/position';
 import { Chunk, ChunkUpdateItem } from '@engine/world/map/chunk';
 import { WorldItem } from '@engine/world/items/world-item';
 import { ByteBuffer } from '@runejs/core/buffer';
-import { Npc } from '@engine/world/actor/npc/npc';
+import { Npc } from '@engine/world/actor/npc';
 import { stringToLong } from '@engine/util/strings';
 import { LandscapeObject } from '@runejs/filestore';
 import { xteaRegions } from '@engine/config/config-handler';
