+++ src/game-engine/util/varbits.ts
@@ -1,7 +1,7 @@
 import { filestore } from '@engine/game-server';
 import { findNpc } from '@engine/config/config-handler';
 import { logger } from '@runejs/core';
-import { Npc } from '@engine/world/actor/npc/npc';
+import { Npc } from '@engine/world/actor/npc';
 import { Player } from '@engine/world/actor/player/player';
 
 const varbitMasks = [];
