+++ src/game-engine/world/action/hooks/task.ts
@@ -2,10 +2,10 @@ import uuidv4 from 'uuid/v4';
 import { lastValueFrom, Subscription, timer } from 'rxjs';
 import { Actor } from '@engine/world/actor/actor';
 import { ActionHook } from '@engine/world/action/action-pipeline';
-import { World } from '@engine/world';
+import { World } from '@engine/world/world';
 import { logger } from '@runejs/core';
 import { Player } from '@engine/world/actor/player/player';
-import { Npc } from '@engine/world/actor/npc/npc';
+import { Npc } from '@engine/world/actor/npc';
 import { ActionStrength } from '@engine/world/action/action-pipeline';
