+++ src/game-engine/world/action/item-on-npc.action.ts
@@ -2,7 +2,7 @@ import { Player } from '@engine/world/actor/player/player';
 import { Position } from '@engine/world/position';
 import { ActionHook, getActionHooks } from '@engine/world/action/hooks';
 import { Item } from '@engine/world/items/item';
-import { Npc } from '@engine/world/actor/npc/npc';
+import { Npc } from '@engine/world/actor/npc';
 import { advancedNumberHookFilter, questHookFilter, stringHookFilter } from '@engine/world/action/hooks/hook-filters';
 import { ActionPipe, RunnableHooks } from '@engine/world/action/action-pipeline';
