+++ src/game-engine/world/action/magic-on-npc.action.ts
@@ -2,7 +2,7 @@ import { Player } from '@engine/world/actor/player/player';
 import { ActionHook, getActionHooks } from '@engine/world/action/hooks';
 import { advancedNumberHookFilter, questHookFilter } from '@engine/world/action/hooks/hook-filters';
 import { ActionPipe, RunnableHooks } from '@engine/world/action/action-pipeline';
-import { Npc } from '../actor/npc/npc';
+import { Npc } from '../actor/npc';
 
 
 /**
