+++ src/game-engine/world/action/npc-init.action.ts
@@ -1,4 +1,4 @@
-import { Npc } from '@engine/world/actor/npc/npc';
+import { Npc } from '@engine/world/actor/npc';
 import { ActionHook, getActionHooks } from '@engine/world/action/hooks';
 import { stringHookFilter } from '@engine/world/action/hooks/hook-filters';
 import { ActionPipe } from '@engine/world/action/action-pipeline';
