+++ src/game-engine/world/actor/behaviors/auto-attack.behavior.ts
@@ -5,7 +5,7 @@ import { logger } from '@runejs/core';
 import { Behavior, BehaviorType } from './behavior';
 import { Timestamp } from 'rxjs/dist/types/internal/types';
 import { timestamp } from 'rxjs/dist/types/operators';
-import { Npc } from '../npc/npc';
+import { Npc } from '../npc';
 import { Player } from '../player/player';
 
 export class AutoAttackBehavior extends Behavior {
