+++ src/game-engine/world/actor/index.ts
@@ -1,0 +1,13 @@
+export * from './actor';
+export * from './player';
+export * from './npc';
+
+export * from './combat';
+export * from './dialogue';
+export * from './effect';
+export * from './magic';
+export * from './pathfinding';
+export * from './prayer';
+export * from './skills';
+export * from './update-flags';
+export * from './walking-queue';
