+++ src/game-engine/world/actor/npc.ts
@@ -3,19 +3,19 @@ import uuidv4 from 'uuid/v4';
 import { Position } from '@engine/world/position';
 import { filestore, world } from '@engine/game-server';
 import { directionData } from '@engine/world/direction';
-import { QuadtreeKey } from '@engine/world';
+import { QuadtreeKey } from '@engine/world/world';
 import { findItem, findNpc } from '@engine/config/config-handler';
 import { animationIds } from '@engine/world/config/animation-ids';
 import { DropTable, NpcCombatAnimations, NpcDetails } from '@engine/config/npc-config';
 import { SkillName } from '@engine/world/actor/skills';
 import { NpcSpawn } from '@engine/config/npc-spawn-config';
-import { MeleeCombatBehavior } from '../behaviors/melee-combat.behavior';
+import { MeleeCombatBehavior } from './behaviors/melee-combat.behavior';
 import { forEach } from 'lodash';
-import { Behavior } from '../behaviors/behavior';
+import { Behavior } from './behaviors/behavior';
 import EventEmitter from 'events';
-import { soundIds } from '../../config/sound-ids';
-import { Player } from '../player/player';
-import { itemIds } from '../../config/item-ids';
+import { soundIds } from '../config/sound-ids';
+import { Player } from './player/player';
+import { itemIds } from '../config/item-ids';
 import { logger } from '@runejs/core';
