+++ src/game-engine/world/actor/player/index.ts
@@ -1,0 +1,10 @@
+export * from './sync';
+export * from './achievements';
+export * from './attack';
+export * from './cutscenes';
+export * from './interface-state';
+export * from './model';
+export * from './player';
+export * from './player-data';
+export * from './private-messaging';
+export * from './quest';
