+++ src/game-engine/world/actor/player/sync/index.ts
@@ -1,0 +1,3 @@
+export * from './actor-sync';
+export * from './npc-sync-task';
+export * from './player-sync-task';
