+++ src/game-engine/world/actor/walking-queue.ts
@@ -2,7 +2,7 @@ import { Actor } from './actor';
 import { Position } from '../position';
 import { Player } from './player/player';
 import { world } from '@engine/game-server';
-import { Npc } from './npc/npc';
+import { Npc } from './npc';
 import { regionChangeActionFactory } from '@engine/world/action/region-change.action';
 import { Subject } from 'rxjs';
