+++ src/game-engine/world/items/index.ts
@@ -1,0 +1,3 @@
+export * from './item';
+export * from './item-container';
+export * from './world-item';
