+++ src/plugins/combat/combat.old.ts
@@ -2,10 +2,10 @@ import { NpcInteractionAction, npcInteractionActionHandler } from '@engine/world
 import { Actor } from '@engine/world/actor/actor';
 import { Player } from '@engine/world/actor/player/player';
 import { lastValueFrom, timer } from 'rxjs';
-import { World } from '@engine/world';
+import { World } from '@engine/world/world';
 import { filter, take } from 'rxjs/operators';
 import { animationIds } from '@engine/world/config/animation-ids';
-import { Npc } from '@engine/world/actor/npc/npc';
+import { Npc } from '@engine/world/actor/npc';
 import { world } from '@engine/game-server';
 import { itemIds } from '@engine/world/config/item-ids';
 import { soundIds } from '@engine/world/config/sound-ids';
