+++ src/plugins/commands/spawn-npc-command.plugin.ts
@@ -1,6 +1,6 @@
 import { commandActionHandler } from '@engine/world/action/player-command.action';
 import { world } from '@engine/game-server';
-import { Npc } from '@engine/world/actor/npc/npc';
+import { Npc } from '@engine/world/actor/npc';
 import { findNpc } from '@engine/config/config-handler';
 import { NpcDetails } from '@engine/config/npc-config';
 import { NpcSpawn } from '@engine/config/npc-spawn-config';
