+++ src/plugins/skills/mining/prospecting.plugin.ts
@@ -1,6 +1,6 @@
 import { objectInteractionActionHandler } from '@engine/world/action/object-interaction.action';
 import { soundIds } from '@engine/world/config/sound-ids';
-import { World } from '@engine/world';
+import { World } from '@engine/world/world';
 import { filestore } from '@engine/game-server';
 import { getAllOreIds, getOreFromRock } from '@engine/world/config/harvestable-object';
 import { findItem } from '@engine/config/config-handler';
